requirements.txt
spotipy
google-genai
streamlit
python-dotenv